V 0.0.1 the module about dicom view has been finished

+++ Controller/Log.py
@@ -17,8 +17,8 @@ def LogTrace(name):
                                                T.tm_hour, T.tm_min, T.tm_sec))
 
 def removeLog():
-
-    os.remove(logname)
+    if os.path.exists(logname):
+        os.remove(logname)
 
 if __name__ == '__main__':
